modified getSharedbyMe function

+++ tests/integration/Owncloud/OcisPhpSdk/ShareCreatedModifyTest.php
@@ -61,9 +61,11 @@ class ShareCreatedModifyTest extends OcisPhpSdkTestCase
         $this->fileToShare->invite([$this->einstein, $this->marie], $this->viewerRole);
         $shares = $this->ocis->getSharedByMe();
         foreach ($shares as $share) {
-            if ($share->getReceiver()->getDisplayName() === 'Albert Einstein') {
-                $share->delete();
-                break;
+            if($share instanceof ShareCreated) {
+                if ($share->getReceiver()->getDisplayName() === 'Albert Einstein') {
+                    $share->delete();
+                    break;
+                }
             }
         }
         $this->assertCount(1, $this->ocis->getSharedByMe());
@@ -83,9 +85,11 @@ class ShareCreatedModifyTest extends OcisPhpSdkTestCase
         $this->fileToShare->invite([$this->einstein, $philosophyHatersGroup, $this->marie], $this->viewerRole);
         $shares = $this->ocis->getSharedByMe();
         foreach ($shares as $share) {
-            if ($share->getReceiver()->getDisplayName() === 'philosophy-haters') {
-                $share->delete();
-                break;
+            if($share instanceof ShareCreated) {
+                if ($share->getReceiver()->getDisplayName() === 'philosophy-haters') {
+                    $share->delete();
+                    break;
+                }
             }
         }
         $this->assertCount(2, $this->ocis->getSharedByMe());
